+++ src/main/java/backend/nomad/dto/store/OrderItemDto.java
@@ -5,10 +5,10 @@ import lombok.Getter;
 @Getter
 public class OrderItemDto {
     private String menuName;
-    private Integer cost;
+    private Double cost;
     private Integer quantity;
 
-    public OrderItemDto(String menuName, Integer cost, Integer quantity) {
+    public OrderItemDto(String menuName, Double cost, Integer quantity) {
         this.menuName = menuName;
         this.cost = cost;
         this.quantity = quantity;
